[MIG] crm_claim_code: Migration to 12.0
[IMP] crm_claim_code: Functions from init was move to hooks

To clear init file

[FIX] End line
