+++ crm_claim_code/README.rst
@@ -1,12 +1,36 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl
-   :alt: License: AGPL-3
+==========================
+Sequential Code for Claims
+==========================
 
-==========================
-Sequential code for claims
-==========================
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fcrm-lightgray.png?logo=github
+    :target: https://github.com/OCA/crm/tree/12.0/crm_claim_code
+    :alt: OCA/crm
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/crm-12-0/crm-12-0-crm_claim_code
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/111/12.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
 
 * This module adds a sequential code for claims.
+
+**Table of contents**
+
+.. contents::
+   :local:
 
 Usage
 =====
@@ -16,24 +40,27 @@ To use this module, you need to:
 * Go to menu **CRM > After Sale > Services > Claims** and create a new claim.
 * Enter claim subject and Save it. You must see a new number for this claim.
 
-
-.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-   :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/111/10.0
-
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/crm/issues>`_. In
-case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smash it by providing detailed and welcomed
-feedback.
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/crm/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/crm/issues/new?body=module:%20crm_claim_code%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
+Authors
+~~~~~~~
+
+* AvanzOSC
+* Tecnativa
+
 Contributors
-------------
+~~~~~~~~~~~~
 
 * Pedro M. Baeza <pedro.baeza@serviciosbaeza.com>
 * Ana Juaristi <anajuarist@avanzosc.es>
@@ -42,17 +69,20 @@ Contributors
 * Alfredo de la Fuente <alfredodelafuente@avanzosc.es>
 * Vicent Cubells <vicent.cubells@tecnativa.com>
 
-Maintainer
-----------
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
 
 .. image:: https://odoo-community.org/logo.png
    :alt: Odoo Community Association
    :target: https://odoo-community.org
 
-This module is maintained by the OCA.
-
 OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit https://odoo-community.org.
+This module is part of the `OCA/crm <https://github.com/OCA/crm/tree/12.0/crm_claim_code>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
