+++ crm_claim_code/__manifest__.py
@@ -5,7 +5,7 @@
 
 {
     "name": "Sequential Code for Claims",
-    "version": "11.0.1.0.0",
+    "version": "12.0.1.0.0",
     "category": "Customer Relationship Management",
     "author": "AvanzOSC, "
               "Tecnativa, "
