+++ crm_claim_code/i18n/es_MX.po
@@ -22,7 +22,7 @@ msgid "Claim"
 msgstr "Reclamo"
 
 #. module: crm_claim_code
-#: model:ir.model.fields,field_description:crm_claim_code.field_crm_claim_code
+#: model:ir.model.fields,field_description:crm_claim_code.field_crm_claim__code
 msgid "Claim Number"
 msgstr ""
