+++ crm_claim_code/readme/CONTRIBUTORS.rst
@@ -1,0 +1,7 @@
+* Pedro M. Baeza <pedro.baeza@serviciosbaeza.com>
+* Ana Juaristi <anajuarist@avanzosc.es>
+* Iker Coranti <ikercoranti@avanzosc.com>
+* Oihane Crucelaegui <oihanecrucelaegi@avanzosc.es>
+* Alfredo de la Fuente <alfredodelafuente@avanzosc.es>
+* Vicent Cubells <vicent.cubells@tecnativa.com>
+
