+++ crm_claim_code/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+* This module adds a sequential code for claims.
