+++ crm_claim_code/readme/USAGE.rst
@@ -1,0 +1,4 @@
+To use this module, you need to:
+
+* Go to menu **CRM > After Sale > Services > Claims** and create a new claim.
+* Enter claim subject and Save it. You must see a new number for this claim.
